feat: inference embedder class with union encode result
